requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
bcrypt==4.1.2
python-dotenv==1.0.1
pyotp==2.9.0
qrcode==7.4.2
redis==5.0.1
Flask-Mail==0.9.1
requests==2.31.0
PyJWT==2.8.0
gunicorn==21.2.0
prometheus-client==0.20.0
sentry-sdk==1.40.0 